+++ CHANGELOG.md
@@ -20,9 +20,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.0.0] - 2026-06-05
+
 ### Removed
 
-- Hear The Music Tree project page: removed GitHub outbound link
+- Hear The Music Tree project page: removed GitHub outbound link and stars badge
 
 ## [0.2.0]